Rewrite the game in typescript
Game is refactored to abstract the browser away from the game logic.

Typescript is introduced (with various fixes and refactors to facilitate that). Not all type errors are resolved, but the core functionality works.

Uses module imports to function. The HTML page includes a script tag with imports to combine the game and browser abstraction.

The config page is combined and hidden (non-functional), that will have to come back in a later commit.